refactor(routes): extract the /drones core routes into drones.js (god-file decomposition)
Core drone routes (list/pause/resume/status/claim + jobs CRUD) moved verbatim from
mycelium.js into server/routes/drones.js, wired via registerDroneRoutes. POST /drones/jobs's
agentWriteLimiter preserved. mycelium.js -326. Route contract unchanged (284 routes),
tests green. Lucy (GLM-5.2), byte-reviewed by m5Max. Follow-up folds /drones/templates,
/drones/profiles, /drones/artifacts into the same module.
